backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
contextlib2==21.6.0
Django==5.1.4
django-cors-headers==4.7.0
django-ninja==1.4.1
google-auth==2.40.3
google-genai==1.19.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
injector==0.22.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.4.0
websockets==15.0.1
